pyproject.toml
[project]
name = "bot2call"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-error-reporting>=1.12.0",
    "google-cloud-firestore>=2.21.0",
    "ipykernel>=6.30.0",
    "python-dotenv>=1.1.1",
]
